+++ src/linalg/LinearAlgebra/eye.c
@@ -10,13 +10,12 @@
 /*
  * Create a identity matrix with size row x column
  */
-void eye(double* A, int row, int column) {
-	// Reset first
-	memset(A, 0, row * column * sizeof(double));
+void eye(float* A, int row, int column) {
+    // Reset first
+    memset(A, 0, row * column * sizeof(float));
 
-	for (int i = 0; i < row; i++) {
-		*A = 1.0;
-		A += row + 1;
-	}
+    for (int i = 0; i < row; i++) {
+        *A = 1.0;
+        A += row + 1;
+    }
 }
-
