+++ src/linalg/LinearAlgebra/maxvector.c
@@ -8,19 +8,20 @@
 #include "declareFunctions.h"
 
 /*
- * Find the maximum value of vector A with size row x 1, return maximum value and the index of that maximum value
+ * Find the maximum value of vector A with size row x 1, return maximum value and the index of that
+ * maximum value
  */
-void maxvector(double* A, int row, double* val, int* index) {
+void maxvector(float* A, int row, float* val, int* index) {
 
-	double past_val = 0; // initial
-	*index = 0;
-	*val = *(A);
+    float past_val = 0; // initial
+    *index = 0;
+    *val = *(A);
 
-	for (int i = 0; i < row; i++) {
-		if (*(A + i) > past_val) {
-			past_val = *(A + i); // Update
-			*val = *(A + i);
-			*index = i; // Remember the index position
-		}
-	}
+    for (int i = 0; i < row; i++) {
+        if (*(A + i) > past_val) {
+            past_val = *(A + i); // Update
+            *val = *(A + i);
+            *index = i; // Remember the index position
+        }
+    }
 }
